Refactor cyral_repository_network_access_policy
